Refactor to create a basic category/package  (javax.time):
    * TimeSource, Instant, Duration
    * Clock, LocalTime, LocalDate, LocalDateTime, OffsetDate, OffsetTime, OffsetDateTime
    * ZonedDateTime, ZoneId, ZoneOffset
    * Period, PeriodUnit, ISOPeriodUnit
    * AmPMofDay, DayOfWeek, MonthOfYear, QuarterOfYear, Year, YearMonth, MonthDay
    * ZoneResolver, ZoneResolvers, DateResolver, DateResolvers

This corresponds to the description of the high-level API in the overview of JSR 310.
For the current users of java.util Date, Calendar, etc. this should be all they need
to get reliable and correct handling of dates and times and it should be easy to
explain how to convert from current classes to the new classes.

For easy incremental adoption, the new Date/time classes need to be usable with
the current text/locale formatting capabilities in java.util.Formatter and java.text
but that is a topic by itself related to integration with existing classes.

Developers that need more control or more visibility into the semantics would
use the low-level API consisting of:

    * DateTimeField, DateTimeFields, DateTimeRule, DateTimeRuleGroup, DateTimeRuleRange
    * Chronology, ISOChronology, ISODateTimeRule, ISOCalendricalRule
    * PeriodField, PeriodFields, PeriodProvider, InstantProvider
    * CalendricalEngine, Calendrical, CalendricalRule, ExtendedCalendricalRule
    * CalendricalMatcher, CalendricalMatchers
    * WeekRules,
    * DateAdjuster, DateAdjusters, TimeAdjuster
    * TaiInstant, UTCInstant, UTCRules

This split is along the lines of the high/low level API descriptions and will make it
easier for developers to start with just the basics in a single package.